Move Hint component and add stories
