+++ src/components/Hint/index.jsx
